ngtcp2_conn_client_handshake: Make stream_id of type uint64_t

+++ tests/ngtcp2_conn_test.c
@@ -2003,9 +2003,8 @@ void test_ngtcp2_conn_client_handshake(void) {
 
   CU_ASSERT(0 == rv);
 
-  spktlen =
-      ngtcp2_conn_client_handshake(conn, buf, sizeof(buf), &datalen, NULL, 0,
-                                   (int64_t)stream_id, 0, null_data, 199, ++t);
+  spktlen = ngtcp2_conn_client_handshake(conn, buf, sizeof(buf), &datalen, NULL,
+                                         0, stream_id, 0, null_data, 199, ++t);
 
   CU_ASSERT(sizeof(buf) == spktlen);
   CU_ASSERT(199 == datalen);
@@ -2019,9 +2018,8 @@ void test_ngtcp2_conn_client_handshake(void) {
 
   CU_ASSERT(0 == rv);
 
-  spktlen =
-      ngtcp2_conn_client_handshake(conn, buf, sizeof(buf), &datalen, NULL, 0,
-                                   (int64_t)stream_id, 1, null_data, 0, +t);
+  spktlen = ngtcp2_conn_client_handshake(conn, buf, sizeof(buf), &datalen, NULL,
+                                         0, stream_id, 1, null_data, 0, +t);
 
   CU_ASSERT(sizeof(buf) == spktlen);
   CU_ASSERT(0 == datalen);
@@ -2039,7 +2037,7 @@ void test_ngtcp2_conn_client_handshake(void) {
       conn, buf,
       NGTCP2_MIN_LONG_HEADERLEN + 1 + conn->dcid.datalen + conn->scid.datalen +
           300,
-      &datalen, NULL, 0, (int64_t)stream_id, 1, null_data, 0, +t);
+      &datalen, NULL, 0, stream_id, 1, null_data, 0, +t);
 
   CU_ASSERT(spktlen > 0);
   CU_ASSERT(-1 == datalen);
